services/recon-python/pyproject.toml
[project]
name = "bonsai3d-recon"
version = "0.1.0"
description = "Bonsai-3D Python reconstruction service"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0,<1.0",
    "uvicorn[standard]>=0.30.0,<1.0",
    "pillow>=10.3.0,<11.0",
    "numpy>=1.26.0,<2.0",
    "scipy>=1.13.0,<2.0",
    "trimesh>=4.3.0,<5.0",
    "scikit-image>=0.23.0,<1.0",
    "rembg>=2.0.57,<3.0",
    "opencv-python-headless>=4.10.0,<5.0",
]

[project.optional-dependencies]
full = [
    "open3d>=0.18.0,<1.0",
]
dev = [
    "pytest>=8.2.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
target-version = "py311"
line-length = 88
